requirements.txt
requests
python-dotenv
pandas
numpy
matplotlib
seaborn
textblob
vaderSentiment
jupyter
